+++ tests/test_bot.py
@@ -5,12 +5,12 @@
 
 import unittest
 
+from cogs.comicsblog import get_comicsblog
 from cogs.getcomics import getcomics_top_link
 from cogs.google import google_top_link
 from cogs.urban import UrbanSearch
 from cogs.youtube import youtube_top_link
 
-from utils.comicsblog import get_comicsblog
 from utils.gif_json import GifJson
